# docs/swagger.yaml
basePath: /api/v1
definitions:
  Entity.Permission:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  Entity.Role:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      name:
        type: string
      permissionsId:
        items:
          type: string
        type: array
      updatedAt:
        type: string
    type: object
...
  Request.CreatePermission:
    properties:
      Name:
        minLength: 3
        type: string
    required:
    - Name
    type: object
  Request.CreateRole:
    properties:
      Name:
        minLength: 3
        type: string
    required:
    - Name
    type: object
...
  Request.UpdateRole:
    properties:
      Name:
        type: string
    required:
    - Name
    type: object
...
  Response.GeneralResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    required:
    - data
    - error
    - message
    type: object
  Response.GetAllRoles:
    properties:
      Role:
        items:
          $ref: '#/definitions/Entity.Role'
        type: array
    type: object
  Response.GetRole:
    properties:
      Role:
        $ref: '#/definitions/Entity.Role'
    type: object
...
  User.LogoutRequest:
    properties:
      token:
        minLength: 3
        type: string
    required:
    - token
    type: object
host: localhost:3000
info:
  contact:
    email: team@supnex.com
    name: Supnex
    url: https://supne.com/
  description: Golang Monolithic Boilerplate
  title: Golang Monolithic Boilerplate
  version: "1.0"
paths:
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: Logout user with access token
      parameters:
      - description: logout user
        in: body
        name: LogoutUserRequest
        required: true
        schema:
          $ref: '#/definitions/User.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: when access token is not valid
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Logout user
      tags:
      - Auth
...
  /permission:
    get:
      consumes:
      - application/json
      description: Get permission
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.Permission'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get permission
      tags:
      - Permission
    post:
      consumes:
      - application/json
      description: Create permission
      parameters:
      - description: Create permission request
        in: body
        name: createPermissionRequest
        required: true
        schema:
          $ref: '#/definitions/Request.CreatePermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.Permission'
              type: object
        "400":
          description: name should have more than 3 character
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create permission
      tags:
      - Permission
...
  /role/:
    get:
      consumes:
      - application/json
      description: Create role
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Response.GetAllRoles'
              type: object
        "400":
          description: create role
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create role
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Create role
      parameters:
      - description: Create role request
        in: body
        name: CreateRoleRequest
        required: true
        schema:
          $ref: '#/definitions/Request.CreateRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.Role'
              type: object
        "400":
          description: create role
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create role
      tags:
      - Role
  /role/{roleId}:
    delete:
      consumes:
      - application/json
      description: Delete role
      parameters:
      - description: delete role with id
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Response.GetRole'
              type: object
        "400":
          description: get role
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete role
      tags:
      - Role
    get:
      consumes:
      - application/json
      description: Get role
      parameters:
      - description: get role with id
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Response.GetRole'
              type: object
        "400":
          description: get role
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get role
      tags:
      - Role
    patch:
      consumes:
      - application/json
      description: Get role
      parameters:
      - description: update role with id
        in: path
        name: roleId
        required: true
        type: string
      - description: update role model
        in: body
        name: UpdateRole
        required: true
        schema:
          $ref: '#/definitions/Request.UpdateRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Response.GetRole'
              type: object
        "400":
          description: get role
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get role
      tags:
      - Role
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"